Move the emotion service folder to /lib
