+++ lambdas/lib/emotions/nrclex_service.go
