+++ lambdas/lib/emotions/nrclex_service_test.go
